// src/components/EffortLessFeesPayment/types.ts
export interface ContentBlockProps {
  icon: string;

  section?: {
    title: string;
    content: string;
    icon: string;
  }[];
  button?: (
    | {
        title: string;
        color?: undefined;
      }
    | {
        title: string;
        color: string;
      }
  )[];
  id: string;
  direction: "left" | "right";
}